+++ src/component/Authentication/LogIn.js
@@ -1,29 +1,28 @@
 import React, { useState } from 'react';
-import { Link,useHistory } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { login, showAlert } from '../actions'
-import '../css/Authentication.css';
-import { HOST } from '../utils/var';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faExclamationCircle } from '@fortawesome/free-solid-svg-icons'
-import {validatUsername, validatePassword} from '../utils/validate';
+import { login, showAlert } from '../../actions';
+import { HOST } from '../../utils/var';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faExclamationCircle } from '@fortawesome/free-solid-svg-icons';
+import {validatUsername, validatePassword} from '../../utils/validate';
+import '../../css/Authentication.css';
 
-const Signup = ({
+const LogIn = ({
   loginUser,
   alartUser
 }) => {
-
   const [userUsername, updateUsername] = useState('');
   const [userPassword, updatePassword] = useState('');
   const [usernameValidation, updateUsernameValidation] = useState({
     valid: true,
     message: ''
   });
+
   const [passwordValidation, updatePasswordValidation] = useState({
     valid: true,
     message: ''
   })
-  const [userRole, updateRole] = useState('user');
 
   const history = useHistory();
 
@@ -40,10 +39,9 @@ const Signup = ({
       const credentials = {
         username: userUsername.toLowerCase(),
         password: userPassword,
-        role: userRole,
       }
-
-      await fetch(`${HOST}/signup`, {
+  
+      await fetch(`${HOST}/login`, {
         method: "POST",
         mode: 'cors',
         headers: {
@@ -52,7 +50,7 @@ const Signup = ({
         },
         body: JSON.stringify(credentials),
       }).then(response => {
-        if (response.status.toString() === '201') {
+        if (response.status.toString() === '200') {
           response.json().then(data => {
             localStorage.setItem('token', data.token);
             localStorage.setItem('username', data.username);
@@ -60,13 +58,13 @@ const Signup = ({
               username: data.username,
               token: data.token
             });
-            alartUser({message: 'signup successful', positiveOutcome: true});
+            alartUser({message: 'login successful', positiveOutcome: true})
             history.push("/");
           })
         } 
         else {
           response.json().then(data => {
-           alartUser({message: data.errors, positiveOutcome: false})
+            alartUser({message: data.errors + '. please try again', positiveOutcome: false})
           })
         }
       }).catch(function(error) {
@@ -78,44 +76,24 @@ const Signup = ({
   return (
     <div className='auth-container container-fluid text-center bg-img'>
       <div className='rounded py-5 auth'>
-        <h1>Create An Account</h1>
+        <h1>Log In</h1>
+        <p>Hello there, log in and start renting cars <span>&#128515;</span></p>
         <form>
           <input type='text' className='rounded mt-2 authInput' value={userUsername.toLowerCase()} onChange={(e) => updateUsername(e.target.value)} placeholder='username' /><br></br>
           <span className={usernameValidation.valid ? 'd-none' : 'validation-message'}><FontAwesomeIcon icon={faExclamationCircle} /> {usernameValidation.message}</span><br></br>
           <input type='password' className='rounded mt-2 authInput' value={userPassword} onChange={(e) => updatePassword(e.target.value)} placeholder='password' /><br></br>
           <span className={passwordValidation.valid ? 'd-none' : 'validation-message'}><FontAwesomeIcon icon={faExclamationCircle} /> {passwordValidation.message}</span><br></br>
-          <div className='mt-3'>
-          <small><em>How do you intend to use this application?</em></small><br></br>
-          <div>
-            <input
-              type="radio"
-              value="user"
-              className='m-2'
-              checked={userRole === 'user'}
-              onChange={e => updateRole(e.target.value)}
-            />
-            <label>I am renting</label>
-          </div>
-          <div>
-            <input
-              type="radio"
-              value="admin"
-              className='m-2'
-              checked={userRole === 'admin'}
-              onChange={e => updateRole(e.target.value)}
-            />
-            <label>I am a registered rental company</label>
-          </div>
-          </div>
-
-
-          <input type='submit' onClick={handleSubmit} value='Sign up' className='rounded my-3 bg-red shadow' />
+          <input type='submit' onClick={handleSubmit} value='Log In' className='rounded my-3 bg-red shadow' />
         </form>
 
+        <div>
+          <Link to={`/`}>Skip login</Link>
+        </div>
+
         <div className='mt-5'>
-          Already have an account?
+          Don't have an account?
           <br></br>
-          <Link to={`/login`}>Log in</Link>
+          <Link to={`/signup`}>Sign up</Link>
         </div>
       </div>
     </div>
@@ -131,4 +109,4 @@ const mapDispatchToProps = dispatch => ({
   }
 });
 
-export default connect(null, mapDispatchToProps)(Signup);
+export default connect(null, mapDispatchToProps)(LogIn);
